+++ cli/cli/helpers/host_machine_directories/host_machine_directories.go
@@ -17,6 +17,8 @@ const (
 
 	metricsUserIDFilename = "metrics-user-id"
 
+	userConsentToSendMetricsElection = "user-consent-to-send-metrics-election"
+
 	// ------------ Names of dirs inside Kurtosis directory --------------
 	engineDataDirname = "engine-data"
 )
@@ -46,6 +48,15 @@ func GetMetricsUserIdFilepath() (string, error) {
 	filepath, err := xdg.DataFile(xdgRelFilepath)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "An error occurred getting the metrics user id filepath from relative path '%v'", xdgRelFilepath)
+	}
+	return filepath, nil
+}
+
+func GetUserConsentToSendMetricsElectionFilepath() (string, error) {
+	xdgRelFilepath := getRelativeFilepathForXDG(userConsentToSendMetricsElection)
+	filepath, err := xdg.ConfigFile(xdgRelFilepath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred getting the user consent to send metrics election filepath from relative path '%v'", xdgRelFilepath)
 	}
 	return filepath, nil
 }